+++ flash/ek/fonts/font_declaration.hpp
@@ -6,10 +6,10 @@
 namespace ek {
 
 struct font_decl_t {
-    std::vector<uint16_t> sizes;
+    uint16_t fontSize;
     std::vector<font_lib::code_range_t> ranges;
-    bool mirror_case = true;
-    bool use_kerning = true;
+    bool mirrorCase = true;
+    bool useKerning = true;
 };
 
 void from_xml(pugi::xml_node node, font_decl_t& font);