+++ scenex/ek/scenex/text/text_drawer.hpp
@@ -26,11 +26,9 @@ public:
     argb32_t textColor = 0xFFFFFFFF_argb;
     float2 alignment;
     float2 offset;
-    float fixedLineHeight = -1.0f;
     float lineSpacing = 0.0f;
     float letterSpacing = 0.0f;
     bool kerning = true;
-    float size = 16.0f;
     asset_t<Font> font;
     asset_t<Font> nativeFont;